move reusable code resource to object logic to ResourceReturnToObject service
